Adding support to start local golang STOMP broker endpoint.
The broker has the same behavior as the Spring SimpleMessageBroker.
The go stomp server supports both TCP and WebSocket connections and should be configured which one to use when started.
The broker implementation provides callbacks for events like new subscriptions to a topic and incomping requests (both public and private, same as the /pub and /pub/queue message handlers in spring).
Adding EventBus.StartFabricEndpoint() and EventBus.StopFabricEndpoint() APIs which can be used to start and stop the fabric stomp endpoint and the old BrokerConnector.StartTCPServer() API is removed.
Modified the existing go-bifrost “demo” and “service” sample commands to use the new stomp server implementation with optional “tcp” param to indicate whether to use TCP or WebSocket endpoint.

NOTE: currently the stomp server doesn’t support STOMP transactions as they are not used by the application fabric clients.