Adding ServiceRegistry for local fabric services
In order to write a new Fabric service one simply needs to implement the FabricService interface and register the service in the ServiceRegistry.
The base functionality which was exposed by the AbstractService and AbstractBase classes in the Bifrost Java, will now be exposed by the FabricServiceCore interface.

TODO: write unit tests once(if :)) this design is approved & finalized.