Adding StoreSyncService
The new service is initialized automatically the first time when a FabricEndpoint is started via the EventBus.StartFabricEndpoint() API. The sync service provides the same functionality as GalacticStoreService in Java Bifrost – synchronizes the content of the local stores with the content of the client stores.

Deleted the MonitorStream and the entire “util” package. As a replacement the following methods are added to the EventBus interface: AddMonitorEventListener(), SendMonitorEvent() and RemoveMonitorEventListener(). With the new implementation, each event can be processed by multiple consumers (the old MonitorStream was sending events to a single consumer).

Adding “-localhost” flag to the “store” sample command, when specified the sample will connect to localhost:8090 instead of the appfabric.vmware.com. Modified the “service” sample command to initialize a “messageOfTheDayStore” bus store so that the “store” sample can be run against the go stomp broker.