Changing the "Response.Created" type to string
In Bifrost java the Response.created field is defined as java.util.Date which by default is serialized as integer timestamp (the number milliseconds since 1/1/1970).
However, Akmal reported that when connecting to Java AF generated app, the "created" field is serialized as date-time string hence the golang unmarshal operation fails.

To avoid such unmarshal errors I’m changing the type of the “created” field to string which can be used to store both integer timestamps and date-time string values.