Added a couple of new convenience methods to core, cleaned up some docs and formatting. (#10)
* Added a couple of new convenience methods to core

core fabric methods for errors added to allow payload and headers to be submitted. Also cleaned up some of the godoc in the service package.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Disabled some of the demo functions

Cleaner documentation and demo code is required to make these useful. demo needs to ne converted to connecting to local broker instead of cloud.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Updated  README to include better instructions.

More examples and samples are required, however the galactic channels sample code has been updated to work against the local broker and not a non-public endpoint.

Signed-off-by: Dave Shanley <dshanley@vmware.com>