Added new Joke Service to plank, AsyncAPI Contract build out and updates to headers. (#24)
* Added new Plank AsyncAPI sample services contract.

Working through ensuring our samples have a concrete contract.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added default headers to all fabric core messages.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Renamed sample.go to ping-pong-service.go

It makes more sence now and aligns with the rest of the sample serivices.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Created new joke-service

Ported from the older samples used by Transport. I will clean up the older ones when we revamp the readme and started up demo code now that we have plank in place.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added new fabric error definition.

This is an RFC7807 compliant problem object. https://tools.ietf.org/html/rfc7807

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Updated joke service to work correctly.

Needed to be bootable and to make the correct API call.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Fixed logging issue

line break is not rendered correctly, removed it.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added joke service to main demo application.

Also added a few more docs.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added simpler code and another test

This is to address a coverage issue mentioned by @jooskim in PR #24

Signed-off-by: Dave Shanley <dshanley@vmware.com>