Added Reply-To support. (#31)
* Adding in new `SetHttpPathPrefixChannelBridge` method to plank

This allows a full passthrough of everything to a channel via HTTP. A firehose method that is super useful for any application wanting to create multiple instances of plank and direct traffic around to different endpoints, but all the time using the same application bus.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added new integration test harness

This is super useful to building suite tests for plank. Testing out services via both async bus and REST becomes pretty simple.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Refactoring ping-pong service to use newer public properties.

Some of the code in here was useful, so it's be ported into models and turned into re-usable code.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added some more docs and OSS header.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Upgraded to latest go-stomp

New version of go-stomp has updated location of models, and v3.0.3 contains a modification I made to allow a new 'reply-to' API to be added. This is supported on RabbitMQ servers and we can also add support to plank for it also.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Removed borked shortcut to replace code used for local dev.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Addressed comments in #31

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Addressed more comments in #31

Added example from helper method to provide a configurable server config, added some tests for the harness also.

Signed-off-by: Dave Shanley <dshanley@vmware.com>