fix: UUID comparison logic bug
UUID is used throughout the entire codebase of Transport and is
especially important in filtering messages based on their destination.
While implementing basic WASM bridge for the bus I realized that UUID
comparison was failing, i.e. returning true for the equality check
between two dififerent UUID instances. Turns out the use of .ID()
method on the UUID object only returns the first 4 bytes of the
underlying 16-byte slice. This means as long as the first 8
hexadecimal characters matched between two UUID instances, they would
come out as equal, and it would therefore mean a significantly higher
chance of UUID collision.. The fix introduced in this PR is to use
the string comparison of the full UUID in places where .ID() was used.

Signed-off-by: Josh Kim <kjosh@vmware.com>